Must Retry Good question
Question Explanation
We're given with a sorted list of integers (might contain duplicates) in ascending order. We've to return whether the given list can be split into one or more subsequences such that each one is a consecutive increasing sequence and has minimum length of 3.

Example:
[1,2,3,3,4,4,5,5] can be split into [1,2,3,4,5] & [3,4,5]. But,
[1,2,3,5,5,6,7] can't be split like [1,2,3,5] & [5,6,7] because 3 and 5 are not consecutive (3+1 != 5).
Intuition
We can observe that there is no restriction on the number of subsequences that we can generate. Hence, we will traverse the given array and try to push the elements in a single subsequence until we encounter a number which is not consecutive to the previous one. In that case, we can initialize another subsequence with the current element as the starting integer.

Edge Cases

If the input array has length less than 3 then it'll not satisfy the condition.
If the list contains duplicates then we've to check on which subsequence to append our current element.
Approach
We'll initialize a vector<pair<int,int>> v to store different subsequences. Each pair v[i] denotes {last element, length} of ith subsequence. The vector will store {nums[0],1} at the beginning.
We'll iterate nums from 1st index and check whether nums[i] is consecutive to the last element.
Case I: If nums[i] is consecutive, then we'll update the pair of vector with the last element as nums[i] and increase the length by 1. It can be visualized as priority queue where the top element has been popped out and updated pair is pushed in.
Case II: If nums[i] comes out to be equal to the last element (i.e. duplicate), we'll iterate our vector from end to check whether nums[i] can be the next consecutive element of some other subsequence. If we find such a pair, then we'll update it's value, else, we're left with no other option than generating a new subsequence. Here comes the benefit of using vector than C++ STL priority_queue that we can traverse the previous elements without popping out the top elements in the queue.
Case III: If none of the conditions above is true, that means nums[i] has to be the starting element of a new subsequence. Hence, we push_back {nums[i],1} in the vector. It can be visualized as pushing a pair in the priority queue.
After the traversal of nums is completed, we'll iterate through the vector and check whether v[i].second (i.e. length of that subsequence) is greater than 3.

Code

class Solution {
public:
    bool isPossible(vector<int>& nums) {
        int n = nums.size();
        
        if (n < 3) {    // Edge Case #1
            return false;
        }
        
        vector<pair<int,int>> v = {{nums[0],1}};
        
        for (int i = 1; i < n; i++) {
            int sz = v.size();
            pair<int,int> end = v[sz-1];
            
            if (nums[i] == end.first+1) {
                v[sz-1].first++;
                v[sz-1].second++;
            }
            else if (nums[i] == end.first) {   // Edge Case #2
                bool found = false;
                for (int k = sz-2; k >= 0; k--) {
                    if (nums[i] == v[k].first+1) {
                        found = true;
                        v[k].first++;
                        v[k].second++;
                        break;
                    }
                }
                if (!found) {
                    v.push_back({nums[i],1});    
                }
            }
            else {
                v.push_back({nums[i],1});
            }
        }
        
        for (auto &it: v) {
            if (it.second < 3) {
                return false;
            }
        }
        return true;
    }
};
TC: O(n) since the time taken to traverse back for edge case #2 is not considerable
SC: O(n) for the vector v to store the subsequences (in worst case it will be n)