Tough one
Whenever you take a variable length vector it will give a memory limit exceed. So take a constant lenth dp vector according to constraints. And there you go, you solution will work fine.
One more thing, that if you assign the dp array to -1 then you will miss the cases in which the actual multiplicant is -1. That is why when you take this value to be INT_MIN, the solution will get past all test cases.